requirements.txt
jira==3.5.1
pyairtable==2.2.1
python-dotenv==1.0.0
click==8.1.7
apscheduler==3.10.4
requests==2.31.0
python-dateutil==2.8.2
